Move interim chain tests to ol.Tile
